requirements.txt
PyGithub>=2.1.1
requests>=2.31.0
gTTS>=2.5.0
edge-tts>=6.1.9
moviepy>=2.0.0
Pillow>=10.0.0
rich>=13.7.0
icrawler>=0.6.6
